+++ packages/types/src/base.ts
@@ -1,5 +1,9 @@
 import { HDWallet } from '@shapeshiftoss/hdwallet-core'
+<<<<<<< HEAD
 import { QuoteFeeData, SignTxInput } from './chain-adapters'
+=======
+import { ethereum, SignTxInput } from './chain-adapters'
+>>>>>>> main
 
 /** Common */
 
@@ -19,8 +23,8 @@ export enum ContractTypes {
 }
 
 export enum ChainTypes {
-  'Ethereum' = 'ethereum',
-  'Bitcoin' = 'bitcoin'
+  Ethereum = 'ethereum',
+  Bitcoin = 'bitcoin'
 }
 
 export enum NetworkTypes {
@@ -159,7 +163,11 @@ export type Quote<T1 extends ChainTypes, T2 extends SwapperType> = {
   slipScore?: string
   txData?: string
   value?: string
+<<<<<<< HEAD
   feeData?: QuoteFeeData<T1, T2>
+=======
+  feeData?: ethereum.QuoteFeeData
+>>>>>>> main
   allowanceContract?: string
   allowanceGrantRequired?: boolean
   slippage?: string