+++ packages/types/src/chain-adapters/index.ts
@@ -1,16 +1,16 @@
 import { HDWallet, BTCInputScriptType, BTCSignTx, ETHSignTx } from '@shapeshiftoss/hdwallet-core'
 import { BIP32Params, ChainTypes, NetworkTypes, SwapperType } from '../base'
 import { ChainAndPlatformSpecific, ChainSpecific } from '../utility'
-import * as Ethereum from './ethereum'
-import * as Bitcoin from './bitcoin'
+import * as ethereum from './ethereum'
+import * as bitcoin from './bitcoin'
 
-export { Bitcoin, Ethereum }
+export { bitcoin, ethereum }
 
 type ChainSpecificAccount<T> = ChainSpecific<
   T,
   {
-    [ChainTypes.Ethereum]: Ethereum.Account
-    [ChainTypes.Bitcoin]: Bitcoin.Account
+    [ChainTypes.Ethereum]: ethereum.Account
+    [ChainTypes.Bitcoin]: bitcoin.Account
   }
 >
 
@@ -25,7 +25,7 @@ export type Account<T extends ChainTypes> = {
 type ChainSpecificTransaction<T> = ChainSpecific<
   T,
   {
-    [ChainTypes.Bitcoin]: Bitcoin.TransactionSpecific
+    [ChainTypes.Bitcoin]: bitcoin.TransactionSpecific
   }
 >
 
@@ -54,7 +54,7 @@ export enum FeeDataKey {
 type ChainSpecificQuoteFeeData<T1, T2> = ChainAndPlatformSpecific<
   T1,
   {
-    [ChainTypes.Ethereum]: Ethereum.QuoteFeeData
+    [ChainTypes.Ethereum]: ethereum.QuoteFeeData
   },
   T2,
   {
@@ -67,7 +67,7 @@ type ChainSpecificQuoteFeeData<T1, T2> = ChainAndPlatformSpecific<
 type ChainSpecificFeeData<T> = ChainSpecific<
   T,
   {
-    [ChainTypes.Ethereum]: Ethereum.FeeData
+    [ChainTypes.Ethereum]: ethereum.FeeData
   }
 >
 
@@ -75,6 +75,14 @@ export type QuoteFeeData<T1 extends ChainTypes, T2 extends SwapperType> = {
   fee: string
   totalFee: string
 } & ChainSpecificQuoteFeeData<T1, T2>
+
+// ChainTypes.Ethereum:
+// feePerUnit = gasPrice
+// feePerTx = estimateGas (estimated transaction cost)
+// feeLimit = gasLimit (max gas willing to pay)
+
+// ChainTypes.Bitcoin:
+// feePerUnit = sats/kbyte
 
 export type FeeData<T extends ChainTypes> = {
   feePerUnit: string
@@ -84,6 +92,38 @@ export type FeeDataEstimate<T extends ChainTypes> = {
   [FeeDataKey.Slow]: FeeData<T>
   [FeeDataKey.Average]: FeeData<T>
   [FeeDataKey.Fast]: FeeData<T>
+}
+
+export type SubscribeTxsInput = {
+  addresses: Array<string>
+}
+
+export type SubscribeTxsMessage<T extends ChainTypes> = {
+  address: string
+  blockHash?: string
+  blockHeight: number
+  blockTime: number
+  chain: T
+  confirmations: number
+  network: NetworkTypes
+  txid: string
+} & TxTransfer<T>
+
+type ChainSpecificTxTransfer<T> = ChainSpecific<
+  T,
+  {
+    [ChainTypes.Ethereum]: ethereum.TxTransfer
+  }
+>
+
+export type TxTransfer<T extends ChainTypes> = {
+  asset: string
+  type: 'send' | 'receive' | 'fee'
+  value: string
+} & ChainSpecificTxTransfer<T>
+
+export type SubscribeError = {
+  message: string
 }
 
 export type TxHistoryResponse<T extends ChainTypes> = {
@@ -108,7 +148,7 @@ export type BuildSendTxInput = {
   fee?: string
   /** Optional param for eth txs indicating what ERC20 is being sent **/
   erc20ContractAddress?: string
-  recipients?: Array<Bitcoin.Recipient>
+  recipients?: Array<bitcoin.Recipient>
   opReturnData?: string
   scriptType?: BTCInputScriptType
   gasLimit?: string
@@ -133,7 +173,7 @@ export type GetAddressInputBase = {
   bip32Params?: BIP32Params
 }
 
-export type GetAddressInput = GetAddressInputBase | Bitcoin.GetAddressInput
+export type GetAddressInput = GetAddressInputBase | bitcoin.GetAddressInput
 
 export type GetFeeDataInput = {
   contractAddress?: string
@@ -152,4 +192,26 @@ export type ValidAddressResult = {
   valid: boolean
   /** Result type of valid address */
   result: ValidAddressResultType
+}
+
+export type ZrxFeeResult = {
+  fast: number
+  instant: number
+  low: number
+  source:
+    | 'ETH_GAS_STATION'
+    | 'ETHERSCAN'
+    | 'ETHERCHAIN'
+    | 'GAS_NOW'
+    | 'MY_CRYPTO'
+    | 'UP_VEST'
+    | 'GETH_PENDING'
+    | 'MEDIAN'
+    | 'AVERAGE'
+  standard: number
+  timestamp: number
+}
+
+export type ZrxGasApiResponse = {
+  result: ZrxFeeResult[]
 }